Fixed the subject->thread code, using a regex.  Following NathanW's lead, I broke it into a new function (subject2thread), but put it in its own file (thread.cpp) and added a test harness (retest.cpp).
The test harness adds main() to Parser making it an executable
(not just an addon); give it a list of subject headers as input and
it generates the corresponding thread name.  Together with NathanW,
I compiled a list of subjects that should stress the threading code
(subjects).  A new makefile target (test) executes Parser with those
subjects as input.